OntologyMapping

Documentation for OntologyMapping.

OntologyMapping.get_ontology_mapping

OntologyMapping.map_to_ontology

OntologyMapping.get_ontology_mapping — Method

get_ontology_mapping(to_map_ids::Vector{String}, ontology::String;
                     field::Symbol=:short_form)

Get a mapping of cell types to an ontology. This performs a search for each cell type in the ontology and returns the mapping.

OntologyMapping.map_to_ontology — Method

map_to_ontology(eset::ExpressionSet, ontology::String;
                ontology_col::String,
                target_col::String="ontology_ids",
                replacements::NTuple{K,Pair{String,String}}) where {K}

Maps cell types in an ExpressionSet to a specified ontology and adds the mapping to the ExpressionSet.

Arguments

eset::ExpressionSet: The expression set containing the data to be mapped.

ontology::String: The ontology to map the cell types to.

ontology_col::String: The column in the phenotype data that contains the cell types to be mapped.

target_col::String="ontology_ids": The column name where the ontology IDs will be stored in the phenotype data. Defaults to "ontology_ids".

replacements::NTuple{K,Pair{String,String}}: A tuple of string pairs for replacing substrings in the cell types to improve mapping accuracy.